fix: gate review approval on central sync and preserve pipeline failure reason
