+++ src/synthorg/engine/task_sync.py
@@ -9,20 +9,18 @@ and no recorded no-op justification is driven to ``FAILED`` rather than
 pushed to review as a silent no-op success.
 """
 
-import asyncio
 from typing import TYPE_CHECKING, Final
-from uuid import uuid4
 
 from synthorg.approval.protocol import ApprovalStoreProtocol
 from synthorg.core.critical_errors import reraise_critical
 from synthorg.core.run_outcome import RunOutcome
 from synthorg.core.task_enums import TaskStatus
+from synthorg.engine._task_sync_engine import sync_to_task_engine
 from synthorg.engine.context import AgentContext
-from synthorg.engine.errors import ExecutionStateError, TaskEngineError
+from synthorg.engine.errors import ExecutionStateError
 from synthorg.engine.loop_protocol import ExecutionResult, TerminationReason
 from synthorg.engine.resume_scope import is_resumed_run
 from synthorg.engine.task_engine import TaskEngine
-from synthorg.engine.task_engine_models import TransitionTaskMutation
 from synthorg.engine.task_sync_review import create_review_approval
 from synthorg.observability import get_logger, safe_error_description
 
@@ -35,8 +33,6 @@ if TYPE_CHECKING:
 from synthorg.observability.events.execution import (
     EXECUTION_ENGINE_ERROR,
     EXECUTION_ENGINE_NO_ARTIFACTS_FAILED,
-    EXECUTION_ENGINE_SYNC_FAILED,
-    EXECUTION_ENGINE_TASK_SYNCED,
     EXECUTION_ENGINE_TASK_TRANSITION,
 )
 
@@ -65,119 +61,6 @@ _EMPTY_RUN_REASON: Final[str] = (
 # it MUST be a system/pipeline-set, validated signal, never a value derived
 # from agent/LLM output, so an agent cannot self-justify an empty run.
 _NO_OP_JUSTIFICATION_KEY: Final[str] = "no_op_justification"
-
-
-async def sync_to_task_engine(  # noqa: PLR0913
-    task_engine: TaskEngine | None,
-    *,
-    target_status: TaskStatus,
-    task_id: str,
-    agent_id: str,
-    reason: str,
-    critical: bool = False,
-) -> bool:
-    """Sync a status transition to the centralized TaskEngine.
-
-    Best-effort: failures are logged and swallowed so that agent
-    execution is never blocked by a TaskEngine issue.
-    ``MemoryError`` and ``RecursionError`` propagate unconditionally.
-
-    Args:
-        task_engine: The task engine to sync to, or ``None`` (no-op).
-        target_status: The status to transition to.
-        task_id: Task identifier.
-        agent_id: Agent performing the transition.
-        reason: Human-readable reason for the transition.
-        critical: If ``True``, sync failure is logged at ERROR level
-            instead of WARNING (severity only -- sync remains best-effort
-            regardless).
-
-    Returns:
-        ``True`` when the central engine now reflects the transition
-        (no engine wired, so the local state is authoritative; or the
-        mutation succeeded); ``False`` when an engine is wired but the
-        mutation was swallowed or rejected, so the central task is still
-        in its prior status. A caller that persists downstream state
-        keyed on the new status (e.g. a failure approval) must gate on
-        this so it never references a status the engine has not applied.
-
-    Raises:
-        MemoryError: Propagated unconditionally (non-recoverable).
-        RecursionError: Propagated unconditionally (non-recoverable).
-        asyncio.CancelledError: Propagated so shutdown can proceed.
-    """
-    if task_engine is None:
-        return True
-
-    try:
-        mutation = TransitionTaskMutation(
-            request_id=uuid4().hex,
-            requested_by=agent_id,
-            task_id=task_id,
-            target_status=target_status,
-            reason=reason,
-        )
-        result = await task_engine.submit(mutation)
-    except asyncio.CancelledError:
-        raise
-    except Exception as exc:  # noqa: BLE001 -- criticals re-raised
-        # lint-allow: swallow-ok -- best-effort side channel
-        reraise_critical(exc)
-        _log_sync_issue(
-            critical=critical,
-            agent_id=agent_id,
-            task_id=task_id,
-            target_status=target_status,
-            error=(
-                "TaskEngine unavailable"
-                if isinstance(exc, TaskEngineError)
-                else "Unexpected error syncing to TaskEngine"
-            ),
-        )
-        return False
-
-    if result.success:
-        logger.debug(
-            EXECUTION_ENGINE_TASK_SYNCED,
-            agent_id=agent_id,
-            task_id=task_id,
-            target_status=target_status.value,
-            version=result.version,
-        )
-        return True
-
-    # Mutation was rejected (e.g. version conflict, invalid
-    # transition, task not found).
-    _log_sync_issue(
-        critical=critical,
-        agent_id=agent_id,
-        task_id=task_id,
-        target_status=target_status,
-        error=result.error or "Mutation rejected (no error detail)",
-        error_code=result.error_code,
-    )
-    return False
-
-
-def _log_sync_issue(
-    *,
-    critical: bool,
-    agent_id: str,
-    task_id: str,
-    target_status: TaskStatus,
-    **extra: object,
-) -> None:
-    """Log a sync failure at ERROR (critical) or WARNING severity."""
-    common = {
-        "agent_id": agent_id,
-        "task_id": task_id,
-        "target_status": target_status.value,
-        **extra,
-    }
-    if critical:
-        logger.error(EXECUTION_ENGINE_SYNC_FAILED, **common)
-    else:
-        logger.warning(EXECUTION_ENGINE_SYNC_FAILED, **common)
 
 
 async def transition_task_if_needed(
@@ -410,9 +293,10 @@ async def _transition_to_review(  # noqa: PLR0913 -- post-exec collaborators
         The original ``execution_result`` when the context is unchanged, or
         a copy carrying the furthest-reached context.
     """
+    synced = False
     for target, step_reason in _COMPLETION_STEPS:
         try:
-            ctx, _ = await _transition_and_sync(
+            ctx, synced = await _transition_and_sync(
                 ctx,
                 target_status=target,
                 reason=step_reason,
@@ -431,8 +315,13 @@ async def _transition_to_review(  # noqa: PLR0913 -- post-exec collaborators
             )
             break
 
+    # Only queue the review approval + auto-review once the central engine
+    # reflects IN_REVIEW: a swallowed/rejected sync leaves the engine's task
+    # IN_PROGRESS, and an approval (or an auto-review decision) against that
+    # stale state would act on a status the engine has not applied.
     if (
-        ctx.task_execution is not None
+        synced
+        and ctx.task_execution is not None
         and ctx.task_execution.status == TaskStatus.IN_REVIEW
     ):
         # Emptiness is an artifact-count fact resolved truthfully at read time,
