+++ tests/unit/engine/test_task_sync.py
@@ -11,6 +11,7 @@ from synthorg.approval.protocol import ApprovalStoreProtocol
 from synthorg.core.artifact import ArtifactType, ExpectedArtifact
 from synthorg.core.completion_enums import FinishReason
 from synthorg.core.task_enums import TaskStatus
+from synthorg.engine._task_sync_engine import sync_to_task_engine
 from synthorg.engine.context import AgentContext
 from synthorg.engine.errors import ExecutionStateError, TaskEngineError
 from synthorg.engine.loop_protocol import (
@@ -27,7 +28,6 @@ from synthorg.engine.task_engine_models import (
 )
 from synthorg.engine.task_sync import (
     apply_post_execution_transitions,
-    sync_to_task_engine,
     transition_task_if_needed,
 )
 from synthorg.engine.task_sync_review import _REVIEW_ACTION_TYPE
@@ -245,7 +245,7 @@ class TestSyncToTaskEngine:
             side_effect=TaskEngineError("unavailable"),
         )
 
-        with patch("synthorg.engine.task_sync.logger") as mock_logger:
+        with patch("synthorg.engine._task_sync_engine.logger") as mock_logger:
             await sync_to_task_engine(
                 mock_te,
                 target_status=TaskStatus.IN_PROGRESS,
@@ -925,6 +925,41 @@ class TestReviewApprovalCreation:
         assert item.action_type == _REVIEW_ACTION_TYPE
         assert item.task_id == str(sample_task_with_criteria.id)
         assert item.status == ApprovalStatus.PENDING
+
+    async def test_skips_approval_when_review_sync_rejected(
+        self,
+        sample_agent_with_personality: AgentIdentity,
+        sample_task_with_criteria: Task,
+    ) -> None:
+        """No review approval is queued when the engine rejects the IN_REVIEW sync.
+
+        A swallowed/rejected central transition leaves the engine's task
+        IN_PROGRESS; an approval against that stale state would let a later
+        decision act on a status the engine has not applied.
+        """
+        ctx = AgentContext.from_identity(
+            sample_agent_with_personality,
+            task=sample_task_with_criteria,
+        )
+        ctx = ctx.with_task_transition(TaskStatus.IN_PROGRESS, reason="started")
+        result = _make_execution_result(ctx, reason=TerminationReason.COMPLETED)
+        mock_store = mock_of[ApprovalStoreProtocol](add=AsyncMock())
+        mock_te = _make_mock_task_engine(return_value=_make_sync_failure())
+
+        out = await apply_post_execution_transitions(
+            result,
+            agent_id=str(sample_agent_with_personality.id),
+            task_id=str(sample_task_with_criteria.id),
+            task_engine=mock_te,
+            approval_store=mock_store,
+        )
+
+        # Local state still reflects IN_REVIEW (the local transition is applied
+        # unconditionally), but the approval is withheld because the engine did
+        # not accept the transition.
+        assert out.context.task_execution is not None
+        assert out.context.task_execution.status == TaskStatus.IN_REVIEW
+        mock_store.add.assert_not_awaited()
 
     async def test_no_approval_without_store(
         self,
